Fix minor code smells found by SonarCloud

+++ src/http_req.c
@@ -145,7 +145,7 @@ http_req_header_get_field(const http_req_header_t req_header, const char* const 
         {
             return NULL;
         }
-        *p_len = (uint32_t)(ptrdiff_t)(p_end - p_val);
+        *p_len = (uint32_t)(p_end - p_val);
     }
     else
     {
@@ -154,7 +154,7 @@ http_req_header_get_field(const http_req_header_t req_header, const char* const 
         {
             return NULL;
         }
-        *p_len = (uint32_t)(ptrdiff_t)(p_end - p_val);
+        *p_len = (uint32_t)(p_end - p_val);
     }
 
     return p_val;
